Implement an actual NonLoadingClassWriter.getCommonSuperClass() method
This should fix most stack frame map bugs; Java 7 is known to be
especially picky about those.

The implementation is based on the official ASM sources, except it
does not load classes (except for blacklisted classes) and hence can
be used safely inside the agent. The routine is far from perfect
though (see TODO in the code).

ContractClassfileTransformer now requires a correct class loader,
which can look up source files as well as general class files that
comprise the class hierarchy---class files are read and cached as
needed by getCommonSuperClass(). (The cache is currently never
pruned, which is probably bad.)

It is possible not to provide a class loader, if the default one know
how to find these resources (which is typically the case when
instrumenting at run time, but may not be during offline
instrumentation; see TODO in PreMain.main()).

The Ant cofojab task has been updated to correctly retrieve and pass
along the class path information through the Ant custom class loader.