Add support for Java 8 lambdas + test
Java 8 lambdas generate synthetic methods used in invokedynamic
instructions. The synthetic methods need to be copied and renamed,
similarly to access methods. Bootstrap method entries need to be
adjusted; in ASM, this is done through the visitInvokeDynamicInsn()
method directly as the ASM class writer takes care of the constant
pool and bootstrap method table for us.